requirements.txt
duckdb==1.1.3
pandas
streamlit
langgraph
pydantic
requests
pypdf
faiss-cpu
sentence-transformers
langchain-text-splitters
python-dotenv
tabulate
google-generativeai
# sqlite3-binary
uuid
matplotlib